move HTMLRenderer to html.py
